+++ lib/models/cache_item.dart
@@ -16,4 +16,22 @@ class CacheItem<T> {
         expiry: DateTime.now().add(slidingExpiry!),
         slidingExpiry: slidingExpiry);
   }
+
+  Map<String, dynamic> toJson() {
+    return {
+      'value': value,
+      'expiry': expiry?.toIso8601String(),
+      'slidingExpiry': slidingExpiry?.inSeconds,
+    };
+  }
+
+  factory CacheItem.fromJson(Map<String, dynamic> json) {
+    return CacheItem<T>(
+      value: json['value'] as T,
+      expiry: json['expiry'] != null ? DateTime.parse(json['expiry']) : null,
+      slidingExpiry: json['slidingExpiry'] != null
+          ? Duration(seconds: json['slidingExpiry'])
+          : null,
+    );
+  }
 }